Update chat_viewer.xml
As a recent convert to Xabber on Android, only one small thing annoyed me about the interface - when I'm chatting with a female user, and she goes offline, the app tells me that messages I send "will be delivered when *he* comes back online". 

Since there's no generally reliable way to tell the gender of a contact, it would be better to use the gender-neutral form "when they come back online" in this message.