feat(MUC): Add Preference to Suppress XEP Status Code 100
XEP Status Code 100 tells the XMPP client application, that a room is not
anonymous, which means that JIDs of all members are revealed.

Unfortunately, XEP states, that the Server should send an body together
with the status code, containing 'This room is not anonymous'.

Extract from XEP 1.24:
"Example 24. Service Warns New Occupant About Lack of Anonymity

<message
    from='darkcave@chat.shakespeare.lit'
    to='hag66@shakespeare.lit/pda'
    type='groupchat'>
  <body>This room is not anonymous.</body>
  <x xmlns='http://jabber.org/protocol/muc#user'>
    <status code='100'/>
  </x>
</message>"

Therefore each time a user enters a non-anonymous room he receives a
message containing "This room is not anonymous".
A problem with this is the fact, that this happens with every reconnect
of the mobile connection. For each room with every internet change (WiFi
to mobile or vice versa, or just unstable internet) one'll get a
notificaton.

To stop this from happening I added an Option to the Chat-Room-Settings,
which allows to suppress this messages. This way users can suppress them
for the rooms they don't need this notificaton.